Version 2.2.1
Fixed a bug where JpegDirectory had tag names for image width and height around the wrong way.  Thanks to Sander Smith for pointing this out.

git-svn-id: https://metadata-extractor.googlecode.com/svn/trunk@7 7bd79e3c-ab28-1fa5-d007-79fa9fde61af

+++ src/com/drew/metadata/jpeg/JpegDirectory.java
@@ -43,8 +43,8 @@ public class JpegDirectory extends Directory {
 
 	static {
         tagNameMap.put(new Integer(TAG_JPEG_DATA_PRECISION), "Data Precision");
-        tagNameMap.put(new Integer(TAG_JPEG_IMAGE_HEIGHT), "Image Width");
-        tagNameMap.put(new Integer(TAG_JPEG_IMAGE_WIDTH), "Image Height");
+        tagNameMap.put(new Integer(TAG_JPEG_IMAGE_WIDTH), "Image Width");
+        tagNameMap.put(new Integer(TAG_JPEG_IMAGE_HEIGHT), "Image Height");
 		tagNameMap.put(new Integer(TAG_JPEG_NUMBER_OF_COMPONENTS), "Number of Components");
 		tagNameMap.put(new Integer(TAG_JPEG_COMPONENT_DATA_1), "Component 1");
 		tagNameMap.put(new Integer(TAG_JPEG_COMPONENT_DATA_2), "Component 2");
